Heroic vs Nonheroic levels
So the numbers always came out funny for the official WoTC NPCs... In the core rulebook it says for damage add 1/2 HEROIC levels to damage. For skills it says add 1/2 CHARACTER levels to skills. This changes the system so that all skills pull the original formula. The damage formulas have been adjusted for only heroic level. Comparing NPC stats to the sheet now everything is calculating correctly.

I'm sure there are oddball NPCs that were rushed in some books but this is the correct way this should be handled.